third_party/blink/web_tests/external/wpt/webrtc-extensions/RTCRtpReceiver-jitterBufferTarget-expected.txt

This is a testharness.js-based test.
[FAIL] audio jitterBufferTarget is null by default
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] audio jitterBufferTarget accepts posititve values
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] audio jitterBufferTarget accepts values up to 4000 milliseconds
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] audio jitterBufferTarget returns last valid value on throw
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] audio jitterBufferTarget allows zero value
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] audio jitterBufferTarget allows to reset value to null
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] video jitterBufferTarget is null by default
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] video jitterBufferTarget accepts posititve values
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] video jitterBufferTarget accepts values up to 4000 milliseconds
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] video jitterBufferTarget returns last valid value
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] video jitterBufferTarget allows zero value
  assert_equals: expected (object) null but got (undefined) undefined
[FAIL] video jitterBufferTarget allows to reset value to null
  assert_equals: expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
